Code refactor
Cleaned up the code so that dualgrid is it's own module, with separate utils module for rendering etc.
